fix: fs imporved fault tolerence
list directory was not returning any error messages, now returns appropriate ENOENT or EIO error.
Validate that a stored/existing handle directory handle can be read before operating on it.
Some times, the user can delete dirs while we hold on to a stale handle.
Now we do a read check before resolving a handle and return an appropriate ENOENT or EIO error.