feat: super fast boot optimizations
Changed extension loading to be non-blocking. Earlier, the project load and
following init used to wait for all extension load to complete. This is unwanted as
extension load can happen in parallel with other boot tasks. This will unblock the ui
and the user can start working on the project files as soon as brackets code is loaded.

Also removed some unused extension loading function. It was part of experiments i did
to load extensions from the file syste. But now we load extensions from virtual server
as if it is vended from a url backed with existing code. So deleting.