chore: move line:ch status to left of statusbar. move git icon out of statusbar
We decided to use bottom of right toolbar to toggle bottom panels. Icons on status bar was not
preferred by users we surveyed. statusbar real estate was also less, so would have needed to hide
icons or provided drop ups with confusing ux if we placed icons on statusbar. Changed colors
 of icons to match other toolbar icons. layout fixes for toolbar selected icons.