fix: option-shift-left and right not selecting words in mac
We assigned reserved mac keyboard shortcut for word selection nav to code folding. release
that shortcut to be unused in all platforms.

Expand and collapse current rewired to `ctrl/cmd-shift-[` and `ctrl/cmd-shift-]` in all
platforms as that shortcut was released from redundant recent navigation flows. Also, it
aligns with `ctrl-[`and `ctrl-]` for indent changing.

Expand and collpase all shortcut was never used in usage stats, so removed it.