fix: cursor not seen if line is empty and no git gutter marker in line in git projects
This was caused by fractional .7em rendering of git gutter. The css pixel approximation will hide
the 1px cursor line in certain window widths, so we should not use fractional em in the gutter.
if was to use a 1em gutter and set a fractional em inset border so that gutter width not fractional

also added hover style so that when user hovers over 1em gutter, it expands to 2 em to have
the user not work with needle threading precision to click git gutter. the green git-added-lines
gutter will not expand as there is currently no action associated with it.